geospatial-analysis/requirements.txt
aiormq==6.7.7
numpy>=1.26.2
loguru>=0.7.0
opencv-python>=4.9.0.80
PyYAML>=5.3.1           # read tracker configs
pandas==2.1.4           # export matrix
gdown>=4.7.1            # google drive model download
GitPython>=3.1.0        # track eval cloning
lapx>=0.5.2
boto3>=1.17.0           # aws s3 model download
redis>=3.5.3      
rq==1.15.1
filterpy>=1.4.5         # OCSORT & DeepOCSORT
ultralytics==8.1.14
jsonschema==4.20.0
structlog
Shapely
flask
metaseg
supervision==0.20.0
leafmap
fastapi
uvicorn
ffmpeg-python
onnxruntime
autodistill-grounding-dino
flask_restful
neptune
aiohttp
aiortc
rtsp==1.1.12
click==8.1.7
onnxruntime==1.16.2
mesop